CHROMIUM: mali: Use correct buffer size for nvmem cell read
KASAN detected that we have a slab-out-of-bounds access, and the cause
is that we cast an 1-byte buffer returned by the "volt-bin" nvmem cell
read to u32 pointer, so any operations to the buffer are out-of-bounds
access.

This fixes the issue by using an u8 pointer for the buffer.

BUG=b:247484087
TEST=USE=kasan emerge-corsola chromeos-kernel-5_15
     No more KASAN slab-out-of-bounds errors

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3907124

+++ drivers/gpu/arm/mali/backend/gpu/mali_kbase_devfreq.c
@@ -466,7 +466,7 @@ static int kbase_devfreq_init_core_mask_table(struct kbase_device *kbdev)
 	struct device_node *opp_node = of_parse_phandle(kbdev->dev->of_node,
 			"operating-points-v2", 0);
 	struct nvmem_cell *cell;
-	u32 *buf;
+	u8 *buf;
 	char opp_microvolt_name[30] = "opp-microvolt";
 	u32 table_mapping[MAX_TABS];
 	int table_count;
@@ -485,7 +485,7 @@ static int kbase_devfreq_init_core_mask_table(struct kbase_device *kbdev)
 		goto skip_volt_bin;
 	}
 
-	buf = (u32 *)nvmem_cell_read(cell, &len);
+	buf = (u8 *)nvmem_cell_read(cell, &len);
 	nvmem_cell_put(cell);
 
 	if (IS_ERR(buf)) {
